Move status schema helpers to schema.py
